# Shrink the store's nameToModule map