--- a/internal/wasm/store_module_list_test.go
+++ b/internal/wasm/store_module_list_test.go
@@ -2,6 +2,7 @@ package wasm
 
 import (
 	"context"
+	"fmt"
 	"testing"
 
 	"github.com/tetratelabs/wazero/internal/testing/require"
@@ -15,6 +16,7 @@ func TestStore_registerModule(t *testing.T) {
 		require.NoError(t, s.registerModule(m1))
 		require.Equal(t, map[string]*ModuleInstance{m1.ModuleName: m1}, s.nameToModule)
 		require.Equal(t, m1, s.moduleList)
+		require.Equal(t, 1, s.nameToModuleCap)
 	})
 
 	t.Run("adds second module", func(t *testing.T) {
@@ -22,6 +24,7 @@ func TestStore_registerModule(t *testing.T) {
 		require.NoError(t, s.registerModule(m2))
 		require.Equal(t, map[string]*ModuleInstance{m1.ModuleName: m1, m2.ModuleName: m2}, s.nameToModule)
 		require.Equal(t, m2, s.moduleList)
+		require.Equal(t, 2, s.nameToModuleCap)
 	})
 
 	t.Run("error on duplicated non anonymous", func(t *testing.T) {
@@ -44,6 +47,7 @@ func TestStore_deleteModule(t *testing.T) {
 		// Leaves the other module alone.
 		require.Equal(t, map[string]*ModuleInstance{m1.ModuleName: m1}, s.nameToModule)
 		require.Equal(t, m1, s.moduleList)
+		require.Equal(t, 0, s.nameToModuleCap)
 	})
 
 	t.Run("ok if missing", func(t *testing.T) {
@@ -55,6 +59,7 @@ func TestStore_deleteModule(t *testing.T) {
 
 		require.Zero(t, len(s.nameToModule))
 		require.Nil(t, s.moduleList)
+		require.Equal(t, 0, s.nameToModuleCap)
 	})
 
 	t.Run("delete middle", func(t *testing.T) {
@@ -112,6 +117,53 @@ func TestStore_AliasModule(t *testing.T) {
 		require.Equal(t, map[string]*ModuleInstance{"m1": m1, "m2": m1}, s.nameToModule)
 		// Doesn't affect module names
 		require.Nil(t, s.moduleList)
+		require.Equal(t, 2, s.nameToModuleCap)
+	})
+}
+
+func TestStore_nameToModuleCap(t *testing.T) {
+	t.Run("nameToModuleCap grows beyond initial cap", func(t *testing.T) {
+		s := newStore()
+		for i := 0; i < 300; i++ {
+			require.NoError(t, s.registerModule(&ModuleInstance{ModuleName: fmt.Sprintf("m%d", i)}))
+		}
+
+		require.Equal(t, 300, s.nameToModuleCap)
+	})
+
+	t.Run("nameToModuleCap shrinks by half the cap", func(t *testing.T) {
+		s := newStore()
+		for i := 0; i < 400; i++ {
+			require.NoError(t, s.registerModule(&ModuleInstance{ModuleName: fmt.Sprintf("m%d", i)}))
+		}
+
+		for i := 0; i < 250; i++ {
+			require.NoError(t, s.deleteModule(s.nameToModule[fmt.Sprintf("m%d", i)]))
+		}
+
+		require.Equal(t, 200, s.nameToModuleCap)
+	})
+
+	t.Run("nameToModuleCap does not shrink below initial size", func(t *testing.T) {
+		s := newStore()
+		for i := 0; i < 400; i++ {
+			require.NoError(t, s.registerModule(&ModuleInstance{ModuleName: fmt.Sprintf("m%d", i)}))
+		}
+
+		for i := 0; i < 350; i++ {
+			require.NoError(t, s.deleteModule(s.nameToModule[fmt.Sprintf("m%d", i)]))
+		}
+
+		require.Equal(t, initialNameToModuleListSize, s.nameToModuleCap)
+	})
+
+	t.Run("nameToModuleCap does not grow when if nameToModule does not grow", func(t *testing.T) {
+		s := newStore()
+		for i := 0; i < 400; i++ {
+			require.NoError(t, s.registerModule(&ModuleInstance{ModuleName: fmt.Sprintf("m%d", i)}))
+			require.NoError(t, s.deleteModule(s.nameToModule[fmt.Sprintf("m%d", i)]))
+			require.Equal(t, 1, s.nameToModuleCap)
+		}
 	})
 }