--- a/internal/wasm/store_module_list_test.go
+++ b/internal/wasm/store_module_list_test.go
@@ -16,7 +16,7 @@ func TestStore_registerModule(t *testing.T) {
 		require.NoError(t, s.registerModule(m1))
 		require.Equal(t, map[string]*ModuleInstance{m1.ModuleName: m1}, s.nameToModule)
 		require.Equal(t, m1, s.moduleList)
-		require.Equal(t, 1, s.nameToModuleCap)
+		require.Equal(t, initialNameToModuleListSize, s.nameToModuleCap)
 	})
 
 	t.Run("adds second module", func(t *testing.T) {
@@ -24,7 +24,7 @@ func TestStore_registerModule(t *testing.T) {
 		require.NoError(t, s.registerModule(m2))
 		require.Equal(t, map[string]*ModuleInstance{m1.ModuleName: m1, m2.ModuleName: m2}, s.nameToModule)
 		require.Equal(t, m2, s.moduleList)
-		require.Equal(t, 2, s.nameToModuleCap)
+		require.Equal(t, initialNameToModuleListSize, s.nameToModuleCap)
 	})
 
 	t.Run("error on duplicated non anonymous", func(t *testing.T) {
@@ -47,7 +47,7 @@ func TestStore_deleteModule(t *testing.T) {
 		// Leaves the other module alone.
 		require.Equal(t, map[string]*ModuleInstance{m1.ModuleName: m1}, s.nameToModule)
 		require.Equal(t, m1, s.moduleList)
-		require.Equal(t, 0, s.nameToModuleCap)
+		require.Equal(t, initialNameToModuleListSize, s.nameToModuleCap)
 	})
 
 	t.Run("ok if missing", func(t *testing.T) {
@@ -59,7 +59,7 @@ func TestStore_deleteModule(t *testing.T) {
 
 		require.Zero(t, len(s.nameToModule))
 		require.Nil(t, s.moduleList)
-		require.Equal(t, 0, s.nameToModuleCap)
+		require.Equal(t, initialNameToModuleListSize, s.nameToModuleCap)
 	})
 
 	t.Run("delete middle", func(t *testing.T) {
@@ -117,7 +117,7 @@ func TestStore_AliasModule(t *testing.T) {
 		require.Equal(t, map[string]*ModuleInstance{"m1": m1, "m2": m1}, s.nameToModule)
 		// Doesn't affect module names
 		require.Nil(t, s.moduleList)
-		require.Equal(t, 2, s.nameToModuleCap)
+		require.Equal(t, initialNameToModuleListSize, s.nameToModuleCap)
 	})
 }
 
@@ -159,10 +159,13 @@ func TestStore_nameToModuleCap(t *testing.T) {
 
 	t.Run("nameToModuleCap does not grow when if nameToModule does not grow", func(t *testing.T) {
 		s := newStore()
-		for i := 0; i < 400; i++ {
+		for i := 0; i < 99; i++ {
 			require.NoError(t, s.registerModule(&ModuleInstance{ModuleName: fmt.Sprintf("m%d", i)}))
-			require.NoError(t, s.deleteModule(s.nameToModule[fmt.Sprintf("m%d", i)]))
-			require.Equal(t, 1, s.nameToModuleCap)
+		}
+		for i := 0; i < 400; i++ {
+			require.NoError(t, s.registerModule(&ModuleInstance{ModuleName: fmt.Sprintf("m%d", i+99)}))
+			require.NoError(t, s.deleteModule(s.nameToModule[fmt.Sprintf("m%d", i+99)]))
+			require.Equal(t, initialNameToModuleListSize, s.nameToModuleCap)
 		}
 	})
 }