Putting this out there. Go does not shrink maps when elements are deleted so a long lived map with lots of elements added and deleted can consume a lot of memory.

This change would track the largest size the nameToNode map has grown to and if that size is more than twice the size it needs to be it gets replaced with a new map. Since we wouldn't want to be constantly resizing small maps I've set a floor and set the initial capacity of the map to that floor. I some what arbitrarily chose 10,000 as the floor.

The scenario this helps with is if a long running application has bursty periods of time when it instantiates a lot of named modules but most of the time has a very low volume of named modules.

I think it might be nice to make an option to disable or adjust the floor for applications running in memory constrained environments who know they will never need to instantiate a lot of named modules.

If there is interest in merging this I'll update the tests and add that option.