@ShivanshVij 2023-03-24
It might also make sense to add a knob disabling this - we want the most throughput possible on the system and a massive preallocated map works for that. 

Not the most efficient use of memory, sure, but you can always buy more ram.