@codefromthecrypt 2023-03-25
FWIW I prefer a better, affordable default value vs a million config switches for something that is unlikely to be set by most people. Config is hugely distracting and I recall some places I worked at you'd have a bunch of weedy config and it was hard to feel the project was relevant as they didn't config important things and fuzzed over minute knobs.

my 2p