@ckaznocha 2023-04-15
Updated to shrink the `nameToModule` to adjust to the changes in #1359. Added tests and reduced the initial cap to 100. No settings added for now. Easer to add later than it is remove premature settings!