@ncruces 2023-04-16
Using `initialNameToModuleListSize` seems absolutely necessary to use as a minimum when reclaiming space (to avoid degenerate allocation patterns). But I wonder if it really is important to set the initial capacity.

Since anonymous modules many runtimes will have… less than a handful names in the map? The fixed cost of 100 is probably low, and if it's beneficial to other use cases, I guess it's fine. But I just wonder if it makes sense.