@ckaznocha 2023-04-16
@ncruces fair point. Removed the initial map size so shrinking shouldn't have much impact until the map grows organically.

@codefromthecrypt benchmarks
```
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero/internal/integration_test/bench
                                       │    old.txt    │               new.txt                │
                                       │    sec/op     │    sec/op     vs base                │
Initialization/interpreter-10             11.31µ ± ∞ ¹   12.33µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Initialization/interpreter-multiple-10   12.064µ ± ∞ ¹   9.467µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Initialization/compiler-10                10.76µ ± ∞ ¹   10.18µ ± ∞ ¹       ~ (p=1.000 n=1) ²
Initialization/compiler-multiple-10       10.48µ ± ∞ ¹   10.50µ ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                   11.14µ         10.57µ        -5.07%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                       │    old.txt    │                new.txt                │
                                       │     B/op      │     B/op       vs base                │
Initialization/interpreter-10            132.3Ki ± ∞ ¹   132.3Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
Initialization/interpreter-multiple-10   132.3Ki ± ∞ ¹   132.3Ki ± ∞ ¹       ~ (p=1.000 n=1) ³
Initialization/compiler-10               132.3Ki ± ∞ ¹   132.3Ki ± ∞ ¹       ~ (p=1.000 n=1) ³
Initialization/compiler-multiple-10      132.3Ki ± ∞ ¹   132.3Ki ± ∞ ¹       ~ (p=1.000 n=1) ³
geomean                                  132.3Ki         132.3Ki        -0.00%
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ need >= 4 samples to detect a difference at alpha level 0.05

                                       │   old.txt   │               new.txt               │
                                       │  allocs/op  │  allocs/op   vs base                │
Initialization/interpreter-10            23.00 ± ∞ ¹   23.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
Initialization/interpreter-multiple-10   23.00 ± ∞ ¹   23.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
Initialization/compiler-10               23.00 ± ∞ ¹   23.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
Initialization/compiler-multiple-10      23.00 ± ∞ ¹   23.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                  23.00         23.00        +0.00%
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
```